requirement.txt
numpy==1.26.4
opencv-python==4.11.0.86
pyyaml==6.0.2
openvino==2023.0.0
pytest==8.4.1
coverage==7.9.1
